requirements.txt
Django>=4.0.0
djangorestframework>=3.13.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0